Split Passes.cpp into separate pass files.
