+++ stablehlo/transforms/MapStablehloToVhlo.h
@@ -38,7 +38,7 @@ struct StablehloToVhloOpImpl {
 template <typename StablehloOpTy>
 using StablehloToVhloOp = typename StablehloToVhloOpImpl<StablehloOpTy>::Type;
 
-#define MAP_STABLEHLO_TO_VERSION(OpName, OpVer)       \
+#define MAP_STABLEHLO_TO_VHLO(OpName, OpVer)       \
   template <>                                         \
   struct StablehloToVhloOpImpl<stablehlo::OpName> {   \
     using Type = vhlo::OpName##OpVer;                 \
@@ -48,125 +48,123 @@ using StablehloToVhloOp = typename StablehloToVhloOpImpl<StablehloOpTy>::Type;
     using Type = stablehlo::OpName;                   \
   };
 
-#define MAP_STABLEHLO_TO_VERSION_V0(OpName) MAP_STABLEHLO_TO_VERSION(OpName, )
+MAP_STABLEHLO_TO_VHLO(AbsOp, V1)
+MAP_STABLEHLO_TO_VHLO(AddOp, V1)
+MAP_STABLEHLO_TO_VHLO(AfterAllOp, V1)
+MAP_STABLEHLO_TO_VHLO(AllGatherOp, V1)
+MAP_STABLEHLO_TO_VHLO(AllReduceOp, V1)
+MAP_STABLEHLO_TO_VHLO(AllToAllOp, V1)
+MAP_STABLEHLO_TO_VHLO(AndOp, V1)
+MAP_STABLEHLO_TO_VHLO(Atan2Op, V1)
+MAP_STABLEHLO_TO_VHLO(BatchNormGradOp, V1)
+MAP_STABLEHLO_TO_VHLO(BatchNormInferenceOp, V1)
+MAP_STABLEHLO_TO_VHLO(BatchNormTrainingOp, V1)
+MAP_STABLEHLO_TO_VHLO(BitcastConvertOp, V1)
+MAP_STABLEHLO_TO_VHLO(BroadcastInDimOp, V1)
+MAP_STABLEHLO_TO_VHLO(BroadcastOp, V1)
+MAP_STABLEHLO_TO_VHLO(CaseOp, V1)
+MAP_STABLEHLO_TO_VHLO(CbrtOp, V1)
+MAP_STABLEHLO_TO_VHLO(CeilOp, V1)
+MAP_STABLEHLO_TO_VHLO(CholeskyOp, V1)
+MAP_STABLEHLO_TO_VHLO(ClampOp, V1)
+MAP_STABLEHLO_TO_VHLO(ClzOp, V1)
+MAP_STABLEHLO_TO_VHLO(CollectivePermuteOp, V1)
+MAP_STABLEHLO_TO_VHLO(CompareOp, V1)
+MAP_STABLEHLO_TO_VHLO(ComplexOp, V1)
+MAP_STABLEHLO_TO_VHLO(ComputeReshapeShapeOp, V1)
+MAP_STABLEHLO_TO_VHLO(ConcatenateOp, V1)
+MAP_STABLEHLO_TO_VHLO(ConstantOp, V1)
+MAP_STABLEHLO_TO_VHLO(ConvertOp, V1)
+MAP_STABLEHLO_TO_VHLO(ConvolutionOp, V1)
+MAP_STABLEHLO_TO_VHLO(CosineOp, V1)
+MAP_STABLEHLO_TO_VHLO(CreateTokenOp, V1)
+MAP_STABLEHLO_TO_VHLO(CrossReplicaSumOp, V1)
+MAP_STABLEHLO_TO_VHLO(CstrReshapableOp, V1)
+MAP_STABLEHLO_TO_VHLO(CustomCallOp, V2)
+MAP_STABLEHLO_TO_VHLO(DivOp, V1)
+MAP_STABLEHLO_TO_VHLO(DotGeneralOp, V1)
+MAP_STABLEHLO_TO_VHLO(DotOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicBroadcastInDimOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicConvOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicGatherOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicIotaOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicPadOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicReshapeOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicSliceOp, V1)
+MAP_STABLEHLO_TO_VHLO(DynamicUpdateSliceOp, V1)
+MAP_STABLEHLO_TO_VHLO(EinsumOp, V1)
+MAP_STABLEHLO_TO_VHLO(Expm1Op, V1)
+MAP_STABLEHLO_TO_VHLO(ExpOp, V1)
+MAP_STABLEHLO_TO_VHLO(FftOp, V1)
+MAP_STABLEHLO_TO_VHLO(FloorOp, V1)
+MAP_STABLEHLO_TO_VHLO(GatherOp, V1)
+MAP_STABLEHLO_TO_VHLO(GetDimensionSizeOp, V1)
+MAP_STABLEHLO_TO_VHLO(GetTupleElementOp, V1)
+MAP_STABLEHLO_TO_VHLO(IfOp, V1)
+MAP_STABLEHLO_TO_VHLO(ImagOp, V1)
+MAP_STABLEHLO_TO_VHLO(InfeedOp, V1)
+MAP_STABLEHLO_TO_VHLO(IotaOp, V1)
+MAP_STABLEHLO_TO_VHLO(IsFiniteOp, V1)
+MAP_STABLEHLO_TO_VHLO(Log1pOp, V1)
+MAP_STABLEHLO_TO_VHLO(LogisticOp, V1)
+MAP_STABLEHLO_TO_VHLO(LogOp, V1)
+MAP_STABLEHLO_TO_VHLO(MapOp, V1)
+MAP_STABLEHLO_TO_VHLO(MaxOp, V1)
+MAP_STABLEHLO_TO_VHLO(MinOp, V1)
+MAP_STABLEHLO_TO_VHLO(MulOp, V1)
+MAP_STABLEHLO_TO_VHLO(NegOp, V1)
+MAP_STABLEHLO_TO_VHLO(NotOp, V1)
+MAP_STABLEHLO_TO_VHLO(OptimizationBarrierOp, V1)
+MAP_STABLEHLO_TO_VHLO(OrOp, V1)
+MAP_STABLEHLO_TO_VHLO(OutfeedOp, V1)
+MAP_STABLEHLO_TO_VHLO(PadOp, V1)
+MAP_STABLEHLO_TO_VHLO(PopulationCountOp, V1)
+MAP_STABLEHLO_TO_VHLO(PowOp, V1)
+MAP_STABLEHLO_TO_VHLO(RealDynamicSliceOp, V1)
+MAP_STABLEHLO_TO_VHLO(RealOp, V1)
+MAP_STABLEHLO_TO_VHLO(RecvOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReduceOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReducePrecisionOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReduceScatterOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReduceWindowOp, V1)
+MAP_STABLEHLO_TO_VHLO(RemOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReplicaIdOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReshapeOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReturnOp, V1)
+MAP_STABLEHLO_TO_VHLO(ReverseOp, V1)
+MAP_STABLEHLO_TO_VHLO(RngBitGeneratorOp, V1)
+MAP_STABLEHLO_TO_VHLO(RngOp, V1)
+MAP_STABLEHLO_TO_VHLO(RoundOp, V1)
+MAP_STABLEHLO_TO_VHLO(RoundNearestEvenOp, V1)
+MAP_STABLEHLO_TO_VHLO(RsqrtOp, V1)
+MAP_STABLEHLO_TO_VHLO(ScatterOp, V1)
+MAP_STABLEHLO_TO_VHLO(SelectAndScatterOp, V1)
+MAP_STABLEHLO_TO_VHLO(SelectOp, V1)
+MAP_STABLEHLO_TO_VHLO(SendOp, V1)
+MAP_STABLEHLO_TO_VHLO(SetDimensionSizeOp, V1)
+MAP_STABLEHLO_TO_VHLO(ShiftLeftOp, V1)
+MAP_STABLEHLO_TO_VHLO(ShiftRightArithmeticOp, V1)
+MAP_STABLEHLO_TO_VHLO(ShiftRightLogicalOp, V1)
+MAP_STABLEHLO_TO_VHLO(SignOp, V1)
+MAP_STABLEHLO_TO_VHLO(SineOp, V1)
+MAP_STABLEHLO_TO_VHLO(SliceOp, V1)
+MAP_STABLEHLO_TO_VHLO(SortOp, V1)
+MAP_STABLEHLO_TO_VHLO(SqrtOp, V1)
+MAP_STABLEHLO_TO_VHLO(SubtractOp, V1)
+MAP_STABLEHLO_TO_VHLO(TanhOp, V1)
+MAP_STABLEHLO_TO_VHLO(TorchIndexSelectOp, V1)
+MAP_STABLEHLO_TO_VHLO(TraceOp, V1)
+MAP_STABLEHLO_TO_VHLO(TransposeOp, V1)
+MAP_STABLEHLO_TO_VHLO(TriangularSolveOp, V1)
+MAP_STABLEHLO_TO_VHLO(TupleOp, V1)
+MAP_STABLEHLO_TO_VHLO(UnaryEinsumOp, V1)
+MAP_STABLEHLO_TO_VHLO(UniformDequantizeOp, V1)
+MAP_STABLEHLO_TO_VHLO(UniformQuantizeOp, V1)
+MAP_STABLEHLO_TO_VHLO(WhileOp, V1)
+MAP_STABLEHLO_TO_VHLO(XorOp, V1)
 
-MAP_STABLEHLO_TO_VERSION_V0(AbsOp)
-MAP_STABLEHLO_TO_VERSION_V0(AddOp)
-MAP_STABLEHLO_TO_VERSION_V0(AfterAllOp)
-MAP_STABLEHLO_TO_VERSION_V0(AllGatherOp)
-MAP_STABLEHLO_TO_VERSION_V0(AllReduceOp)
-MAP_STABLEHLO_TO_VERSION_V0(AllToAllOp)
-MAP_STABLEHLO_TO_VERSION_V0(AndOp)
-MAP_STABLEHLO_TO_VERSION_V0(Atan2Op)
-MAP_STABLEHLO_TO_VERSION_V0(BatchNormGradOp)
-MAP_STABLEHLO_TO_VERSION_V0(BatchNormInferenceOp)
-MAP_STABLEHLO_TO_VERSION_V0(BatchNormTrainingOp)
-MAP_STABLEHLO_TO_VERSION_V0(BitcastConvertOp)
-MAP_STABLEHLO_TO_VERSION_V0(BroadcastInDimOp)
-MAP_STABLEHLO_TO_VERSION_V0(BroadcastOp)
-MAP_STABLEHLO_TO_VERSION_V0(CaseOp)
-MAP_STABLEHLO_TO_VERSION_V0(CbrtOp)
-MAP_STABLEHLO_TO_VERSION_V0(CeilOp)
-MAP_STABLEHLO_TO_VERSION_V0(CholeskyOp)
-MAP_STABLEHLO_TO_VERSION_V0(ClampOp)
-MAP_STABLEHLO_TO_VERSION_V0(ClzOp)
-MAP_STABLEHLO_TO_VERSION_V0(CollectivePermuteOp)
-MAP_STABLEHLO_TO_VERSION_V0(CompareOp)
-MAP_STABLEHLO_TO_VERSION_V0(ComplexOp)
-MAP_STABLEHLO_TO_VERSION_V0(ComputeReshapeShapeOp)
-MAP_STABLEHLO_TO_VERSION_V0(ConcatenateOp)
-MAP_STABLEHLO_TO_VERSION_V0(ConstantOp)
-MAP_STABLEHLO_TO_VERSION_V0(ConvertOp)
-MAP_STABLEHLO_TO_VERSION_V0(ConvolutionOp)
-MAP_STABLEHLO_TO_VERSION_V0(CosineOp)
-MAP_STABLEHLO_TO_VERSION_V0(CreateTokenOp)
-MAP_STABLEHLO_TO_VERSION_V0(CrossReplicaSumOp)
-MAP_STABLEHLO_TO_VERSION_V0(CstrReshapableOp)
-MAP_STABLEHLO_TO_VERSION(CustomCallOp, V2)
-MAP_STABLEHLO_TO_VERSION_V0(DivOp)
-MAP_STABLEHLO_TO_VERSION_V0(DotGeneralOp)
-MAP_STABLEHLO_TO_VERSION_V0(DotOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicBroadcastInDimOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicConvOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicGatherOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicIotaOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicPadOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicReshapeOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicSliceOp)
-MAP_STABLEHLO_TO_VERSION_V0(DynamicUpdateSliceOp)
-MAP_STABLEHLO_TO_VERSION_V0(EinsumOp)
-MAP_STABLEHLO_TO_VERSION_V0(Expm1Op)
-MAP_STABLEHLO_TO_VERSION_V0(ExpOp)
-MAP_STABLEHLO_TO_VERSION_V0(FftOp)
-MAP_STABLEHLO_TO_VERSION_V0(FloorOp)
-MAP_STABLEHLO_TO_VERSION_V0(GatherOp)
-MAP_STABLEHLO_TO_VERSION_V0(GetDimensionSizeOp)
-MAP_STABLEHLO_TO_VERSION_V0(GetTupleElementOp)
-MAP_STABLEHLO_TO_VERSION_V0(IfOp)
-MAP_STABLEHLO_TO_VERSION_V0(ImagOp)
-MAP_STABLEHLO_TO_VERSION_V0(InfeedOp)
-MAP_STABLEHLO_TO_VERSION_V0(IotaOp)
-MAP_STABLEHLO_TO_VERSION_V0(IsFiniteOp)
-MAP_STABLEHLO_TO_VERSION_V0(Log1pOp)
-MAP_STABLEHLO_TO_VERSION_V0(LogisticOp)
-MAP_STABLEHLO_TO_VERSION_V0(LogOp)
-MAP_STABLEHLO_TO_VERSION_V0(MapOp)
-MAP_STABLEHLO_TO_VERSION_V0(MaxOp)
-MAP_STABLEHLO_TO_VERSION_V0(MinOp)
-MAP_STABLEHLO_TO_VERSION_V0(MulOp)
-MAP_STABLEHLO_TO_VERSION_V0(NegOp)
-MAP_STABLEHLO_TO_VERSION_V0(NotOp)
-MAP_STABLEHLO_TO_VERSION_V0(OptimizationBarrierOp)
-MAP_STABLEHLO_TO_VERSION_V0(OrOp)
-MAP_STABLEHLO_TO_VERSION_V0(OutfeedOp)
-MAP_STABLEHLO_TO_VERSION_V0(PadOp)
-MAP_STABLEHLO_TO_VERSION_V0(PopulationCountOp)
-MAP_STABLEHLO_TO_VERSION_V0(PowOp)
-MAP_STABLEHLO_TO_VERSION_V0(RealDynamicSliceOp)
-MAP_STABLEHLO_TO_VERSION_V0(RealOp)
-MAP_STABLEHLO_TO_VERSION_V0(RecvOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReduceOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReducePrecisionOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReduceScatterOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReduceWindowOp)
-MAP_STABLEHLO_TO_VERSION_V0(RemOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReplicaIdOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReshapeOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReturnOp)
-MAP_STABLEHLO_TO_VERSION_V0(ReverseOp)
-MAP_STABLEHLO_TO_VERSION_V0(RngBitGeneratorOp)
-MAP_STABLEHLO_TO_VERSION_V0(RngOp)
-MAP_STABLEHLO_TO_VERSION_V0(RoundOp)
-MAP_STABLEHLO_TO_VERSION_V0(RoundNearestEvenOp)
-MAP_STABLEHLO_TO_VERSION_V0(RsqrtOp)
-MAP_STABLEHLO_TO_VERSION_V0(ScatterOp)
-MAP_STABLEHLO_TO_VERSION_V0(SelectAndScatterOp)
-MAP_STABLEHLO_TO_VERSION_V0(SelectOp)
-MAP_STABLEHLO_TO_VERSION_V0(SendOp)
-MAP_STABLEHLO_TO_VERSION_V0(SetDimensionSizeOp)
-MAP_STABLEHLO_TO_VERSION_V0(ShiftLeftOp)
-MAP_STABLEHLO_TO_VERSION_V0(ShiftRightArithmeticOp)
-MAP_STABLEHLO_TO_VERSION_V0(ShiftRightLogicalOp)
-MAP_STABLEHLO_TO_VERSION_V0(SignOp)
-MAP_STABLEHLO_TO_VERSION_V0(SineOp)
-MAP_STABLEHLO_TO_VERSION_V0(SliceOp)
-MAP_STABLEHLO_TO_VERSION_V0(SortOp)
-MAP_STABLEHLO_TO_VERSION_V0(SqrtOp)
-MAP_STABLEHLO_TO_VERSION_V0(SubtractOp)
-MAP_STABLEHLO_TO_VERSION_V0(TanhOp)
-MAP_STABLEHLO_TO_VERSION_V0(TorchIndexSelectOp)
-MAP_STABLEHLO_TO_VERSION_V0(TraceOp)
-MAP_STABLEHLO_TO_VERSION_V0(TransposeOp)
-MAP_STABLEHLO_TO_VERSION_V0(TriangularSolveOp)
-MAP_STABLEHLO_TO_VERSION_V0(TupleOp)
-MAP_STABLEHLO_TO_VERSION_V0(UnaryEinsumOp)
-MAP_STABLEHLO_TO_VERSION_V0(UniformDequantizeOp)
-MAP_STABLEHLO_TO_VERSION_V0(UniformQuantizeOp)
-MAP_STABLEHLO_TO_VERSION_V0(WhileOp)
-MAP_STABLEHLO_TO_VERSION_V0(XorOp)
-
-#undef MAP_STABLEHLO_TO_VERSION
-#undef MAP_STABLEHLO_TO_VERSION_V0
+#undef MAP_STABLEHLO_TO_VHLO
+#undef MAP_STABLEHLO_TO_VHLO_V0
 
 }  // namespace vhlo
 }  // namespace mlir
